my_book/backend/requirements.txt
fastapi
uvicorn
pinecone
google-generativeai
python-dotenv
openai
pydantic
